normal resultat

+++ src/firstCours/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
     public static void main(String[] args) {
    double nbr1=10,nbr2=3;
-   int resultat= (int) (nbr1/nbr2);
+   double resultat=  (nbr1/nbr2);
    System.out.println(resultat);
     }
 }
